Improve game results display
